HAAR-1933: refactor PUT '/clients/client-credentials/{clientId}/update' to 'base-clients/{baseClientId}'
