+++ src/test/kotlin/uk/gov/justice/digital/hmpps/authorizationserver/integration/ClientsControllerIntTest.kt
@@ -743,6 +743,218 @@ class ClientsControllerIntTest : IntegrationTestBase() {
           CoreMatchers.hasItems("clientId must not be blank"),
         )
     }
+
+    @Nested
+    inner class EditClient {
+
+      @Test
+      fun `access forbidden when no authority`() {
+        webTestClient.put().uri("/base-clients/testy")
+          .body(
+            BodyInserters.fromValue(
+              mapOf(
+                "scopes" to listOf("read", "write"),
+                "authorities" to listOf("CURIOUS_API", "VIEW_PRISONER_DATA", "COMMUNITY"),
+                "ips" to listOf("81.134.202.29/32", "35.176.93.186/32"),
+                "databaseUserName" to "testy-mctest",
+                "jiraNumber" to "HAAR-9999",
+                "validDays" to 5,
+                "accessTokenValidityMinutes" to 20,
+              ),
+            ),
+          )
+          .exchange()
+          .expectStatus().isForbidden
+      }
+
+      @Test
+      fun `access forbidden when no role`() {
+        webTestClient.put().uri("/base-clients/testy")
+          .headers(setAuthorisation(roles = listOf()))
+          .body(
+            BodyInserters.fromValue(
+              mapOf(
+                "scopes" to listOf("read", "write"),
+                "authorities" to listOf("CURIOUS_API", "VIEW_PRISONER_DATA", "COMMUNITY"),
+                "ips" to listOf("81.134.202.29/32", "35.176.93.186/32"),
+                "databaseUserName" to "testy-mctest",
+                "jiraNumber" to "HAAR-9999",
+                "validDays" to 5,
+                "accessTokenValidityMinutes" to 20,
+              ),
+            ),
+          )
+          .exchange()
+          .expectStatus().isForbidden
+      }
+
+      @Test
+      fun `access forbidden when wrong role`() {
+        webTestClient.put().uri("/base-clients/testy")
+          .headers(setAuthorisation(roles = listOf("WRONG")))
+          .body(
+            BodyInserters.fromValue(
+              mapOf(
+                "scopes" to listOf("read", "write"),
+                "authorities" to listOf("CURIOUS_API", "VIEW_PRISONER_DATA", "COMMUNITY"),
+                "ips" to listOf("81.134.202.29/32", "35.176.93.186/32"),
+                "databaseUserName" to "testy-mctest",
+                "jiraNumber" to "HAAR-9999",
+                "validDays" to 5,
+                "accessTokenValidityMinutes" to 20,
+              ),
+            ),
+          )
+          .exchange()
+          .expectStatus().isForbidden
+      }
+
+      @Test
+      fun `not found when client not found`() {
+        webTestClient.put().uri("/base-clients/not-found")
+          .headers(setAuthorisation(roles = listOf("ROLE_OAUTH_ADMIN")))
+          .body(
+            BodyInserters.fromValue(
+              mapOf(
+                "scopes" to listOf("read", "write"),
+                "authorities" to listOf("CURIOUS_API", "VIEW_PRISONER_DATA", "COMMUNITY"),
+                "ips" to listOf("81.134.202.29/32", "35.176.93.186/32"),
+                "databaseUserName" to "testy-mctest",
+                "jiraNumber" to "HAAR-9999",
+                "validDays" to 5,
+                "accessTokenValidityMinutes" to 20,
+              ),
+            ),
+          )
+          .exchange()
+          .expectStatus().isNotFound
+      }
+
+      @Test
+      fun `update complete client success`() {
+        whenever(oAuthClientSecretGenerator.generate()).thenReturn("external-client-secret")
+        whenever(oAuthClientSecretGenerator.encode("external-client-secret")).thenReturn("encoded-client-secret")
+
+        webTestClient.post().uri("/base-clients")
+          .headers(setAuthorisation(roles = listOf("ROLE_OAUTH_ADMIN")))
+          .body(
+            BodyInserters.fromValue(
+              mapOf(
+                "clientId" to "test-test",
+                "clientName" to "testing testing",
+                "scopes" to listOf("read", "write"),
+                "authorities" to listOf("CURIOUS_API", "VIEW_PRISONER_DATA", "COMMUNITY"),
+                "ips" to listOf("81.134.202.29/32", "35.176.93.186/32"),
+                "databaseUserName" to "testy-mctest-1",
+                "jiraNumber" to "HAAR-9999",
+                "validDays" to 5,
+                "accessTokenValidityMinutes" to 20,
+              ),
+            ),
+          )
+          .exchange()
+          .expectStatus().isOk
+
+        webTestClient.put().uri("/base-clients/test-test")
+          .headers(setAuthorisation(roles = listOf("ROLE_OAUTH_ADMIN")))
+          .body(
+            BodyInserters.fromValue(
+              mapOf(
+                "scopes" to listOf("read"),
+                "authorities" to listOf("VIEW_PRISONER_DATA", "COMMUNITY"),
+                "ips" to listOf("82.135.209.29/32", "36.177.94.187/32"),
+                "databaseUserName" to "testy-mctest-2",
+                "jiraNumber" to "HAAR-8888",
+                "validDays" to 3,
+                "accessTokenValidityMinutes" to 10,
+              ),
+            ),
+          )
+          .exchange()
+          .expectStatus().isOk
+
+        val client = clientRepository.findClientByClientId("test-test")
+        assertThat(client!!.scopes).contains("read")
+        assertThat(client.tokenSettings.accessTokenTimeToLive).isEqualTo(Duration.ofMinutes(10))
+        assertThat(client.tokenSettings.settings[RegisteredClientAdditionalInformation.DATABASE_USER_NAME_KEY]).isEqualTo("testy-mctest-2")
+        assertThat(client.tokenSettings.settings[RegisteredClientAdditionalInformation.JIRA_NUMBER_KEY]).isEqualTo("HAAR-8888")
+
+        val clientConfig = clientConfigRepository.findById(client.clientId).get()
+        assertThat(clientConfig.ips).contains("82.135.209.29/32", "36.177.94.187/32")
+        assertThat(clientConfig.clientEndDate).isEqualTo(LocalDate.now().plusDays(2))
+        val authorizationConsent =
+          verifyAuthorities(client.id!!, client.clientId, "ROLE_VIEW_PRISONER_DATA", "ROLE_COMMUNITY")
+
+        verify(telemetryClient).trackEvent(
+          "AuthorizationServerCredentialsUpdate",
+          mapOf("username" to "AUTH_ADM", "clientId" to "test-test"),
+          null,
+        )
+
+        clientRepository.delete(client)
+        clientConfigRepository.delete(clientConfig)
+        authorizationConsentRepository.delete(authorizationConsent)
+      }
+
+      @Test
+      fun `update incomplete client success`() {
+        whenever(oAuthClientSecretGenerator.generate()).thenReturn("external-client-secret")
+        whenever(oAuthClientSecretGenerator.encode("external-client-secret")).thenReturn("encoded-client-secret")
+
+        webTestClient.post().uri("/base-clients")
+          .headers(setAuthorisation(roles = listOf("ROLE_OAUTH_ADMIN")))
+          .body(
+            BodyInserters.fromValue(
+              mapOf(
+                "clientId" to "test-test",
+                "clientName" to "testing testing",
+              ),
+            ),
+          )
+          .exchange()
+          .expectStatus().isOk
+
+        webTestClient.put().uri("/base-clients/test-test")
+          .headers(setAuthorisation(roles = listOf("ROLE_OAUTH_ADMIN")))
+          .body(
+            BodyInserters.fromValue(
+              mapOf(
+                "scopes" to listOf("write"),
+                "authorities" to listOf("VIEW_PRISONER_DATA", "COMMUNITY"),
+                "ips" to listOf("82.135.209.29/32", "36.177.94.187/32"),
+                "databaseUserName" to "testy-mctest-2",
+                "jiraNumber" to "HAAR-8888",
+                "validDays" to 3,
+                "accessTokenValidityMinutes" to 10,
+              ),
+            ),
+          )
+          .exchange()
+          .expectStatus().isOk
+
+        val client = clientRepository.findClientByClientId("test-test")
+        assertThat(client!!.scopes).contains("write")
+        assertThat(client.tokenSettings.accessTokenTimeToLive).isEqualTo(Duration.ofMinutes(10))
+        assertThat(client.tokenSettings.settings[RegisteredClientAdditionalInformation.DATABASE_USER_NAME_KEY]).isEqualTo("testy-mctest-2")
+        assertThat(client.tokenSettings.settings[RegisteredClientAdditionalInformation.JIRA_NUMBER_KEY]).isEqualTo("HAAR-8888")
+
+        val clientConfig = clientConfigRepository.findById(client.clientId).get()
+        assertThat(clientConfig.ips).contains("82.135.209.29/32", "36.177.94.187/32")
+        assertThat(clientConfig.clientEndDate).isEqualTo(LocalDate.now().plusDays(2))
+        val authorizationConsent =
+          verifyAuthorities(client.id!!, client.clientId, "ROLE_VIEW_PRISONER_DATA", "ROLE_COMMUNITY")
+
+        verify(telemetryClient).trackEvent(
+          "AuthorizationServerCredentialsUpdate",
+          mapOf("username" to "AUTH_ADM", "clientId" to "test-test"),
+          null,
+        )
+
+        clientRepository.delete(client)
+        clientConfigRepository.delete(clientConfig)
+        authorizationConsentRepository.delete(authorizationConsent)
+      }
+    }
   }
   private fun verifyAuthorities(id: String, clientId: String, vararg authorities: String): AuthorizationConsent {
     val authorizationConsent = authorizationConsentRepository.findById(AuthorizationConsent.AuthorizationConsentId(id, clientId)).get()
